backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Tetsu.Kuribayashi <tktkchestnut[gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
jupyter = "^1.0.0"
ipykernel = "^6.29.5"
pypdf2 = "^3.0.1"
pdfplumber = "^0.11.3"
openai = "^1.40.3"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
python-multipart = "^0.0.9"
gunicorn = "^23.0.0"
openpyxl = "^3.1.5"
pandas = "^2.2.2"
firebase-admin = "^6.5.0"
python-jose = "^3.3.0"
pydantic = {extras = ["email"], version = "^2.9.2"}
faiss-cpu = "^1.8.0.post1"
scikit-learn = "^1.5.2"
langchain = "^0.3.1"
langchain-community = "^0.3.1"
tiktoken = "^0.7.0"
weaviate-client = "^4.8.1"
pdf2image = "^1.17.0"
pymupdf = "^1.24.11"
inflection = "^0.5.1"
beautifulsoup4 = "^4.12.3"
pydantic-settings = "^2.6.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 88
known_third_party = [
]

[tool.flake8]
max-line-length = 88
exclude = ["build/", "dist/", ".venv/"]  # 除外パス
